+++ src/main/java/com/strangegrotto/wealthdraft/networth/projections/AssetParameterChange.java
@@ -1,21 +1,20 @@
 package com.strangegrotto.wealthdraft.networth.projections;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.google.common.annotations.VisibleForTesting;
 import com.strangegrotto.wealthdraft.errors.ValOrGerr;
 import org.immutables.value.Value;
 
-// TODO Upgrade this to support BigDecimal
+import java.math.BigDecimal;
+
 @Value.Immutable
-@JsonDeserialize(using = AssetParameterChangeDeserializer.class)
 public interface AssetParameterChange {
-    long getValue();
+    BigDecimal getValue();
 
     AssetParameterChangeValueOperation getOperation();
 
     @Value.Derived
-    default ValOrGerr<Long> apply(long oldValue) {
-        ValOrGerr<Long> result = getOperation().apply(oldValue, getValue());
+    default ValOrGerr<BigDecimal> apply(BigDecimal oldValue) {
+        ValOrGerr<BigDecimal> result = getOperation().apply(oldValue, getValue());
         if (result.hasGerr()) {
             return ValOrGerr.propGerr(
                     result.getGerr(),