+++ src/test/java/com/strangegrotto/wealthdraft/networth/projections/AssetParameterChangeDeserializerTest.java
@@ -2,49 +2,41 @@ package com.strangegrotto.wealthdraft.networth.projections;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.strangegrotto.wealthdraft.Main;
 import org.junit.Assert;
 import org.junit.Test;
 
+import java.io.IOException;
+import java.math.BigDecimal;
+
 public class AssetParameterChangeDeserializerTest {
     @Test
-    public void testAddDeserialization() {
-        var mapper = new ObjectMapper();
+    public void testAddDeserialization() throws IOException {
+        var mapper = Main.getObjectMapper();
         var value = 15000L;
         AssetParameterChange change = null;
-        try {
-            change = mapper.readValue("+" + value, AssetParameterChange.class);
-        } catch (JsonProcessingException e) {
-            Assert.fail("No exception should have been thrown on parsing");
-        }
-        Assert.assertEquals(value, change.value);
-        Assert.assertEquals(AssetParameterChangeValueOperation.ADD, change.operation);
+        change = mapper.readValue("+" + value, AssetParameterChange.class);
+        Assert.assertEquals(new BigDecimal(value), change.getValue());
+        Assert.assertEquals(AssetParameterChangeValueOperation.ADD, change.getOperation());
     }
 
     @Test
-    public void testSubtractDeserialization() {
-        var mapper = new ObjectMapper();
+    public void testSubtractDeserialization() throws IOException {
+        var mapper = Main.getObjectMapper();
         var value = 15000L;
         AssetParameterChange change = null;
-        try {
-            change = mapper.readValue("-" + value, AssetParameterChange.class);
-        } catch (JsonProcessingException e) {
-            Assert.fail("No exception should have been thrown on parsing");
-        }
-        Assert.assertEquals(value, change.value);
-        Assert.assertEquals(AssetParameterChangeValueOperation.SUBTRACT, change.operation);
+        change = mapper.readValue("-" + value, AssetParameterChange.class);
+        Assert.assertEquals(new BigDecimal(value), change.getValue());
+        Assert.assertEquals(AssetParameterChangeValueOperation.SUBTRACT, change.getOperation());
     }
 
     @Test
-    public void testSetDeserialization() {
-        var mapper = new ObjectMapper();
+    public void testSetDeserialization() throws IOException {
+        var mapper = Main.getObjectMapper();
         var value = 15000L;
         AssetParameterChange change = null;
-        try {
-            change = mapper.readValue(Long.toString(value), AssetParameterChange.class);
-        } catch (JsonProcessingException e) {
-            Assert.fail("No exception should have been thrown on parsing");
-        }
-        Assert.assertEquals(value, change.value);
-        Assert.assertEquals(AssetParameterChangeValueOperation.SET, change.operation);
+        change = mapper.readValue(Long.toString(value), AssetParameterChange.class);
+        Assert.assertEquals(new BigDecimal(value), change.getValue());
+        Assert.assertEquals(AssetParameterChangeValueOperation.SET, change.getOperation());
     }
 }