+++ src/test/java/com/strangegrotto/wealthdraft/networth/projections/AssetParameterChangeValueOperationTest.java
@@ -3,42 +3,43 @@ package com.strangegrotto.wealthdraft.networth.projections;
 import org.junit.Assert;
 import org.junit.Test;
 
+import java.math.BigDecimal;
+
 public class AssetParameterChangeValueOperationTest {
     @Test
     public void testAdd() {
-        var oldValue = 10L;
-        var toAdd = 5L;
-        var expectedResult = oldValue + toAdd;
+        var oldValue = new BigDecimal(10);
+        var toAdd = new BigDecimal(5);
+        var expectedResult = oldValue.add(toAdd);
         var resultOrErr = AssetParameterChangeValueOperation.ADD.apply(oldValue, toAdd);
         Assert.assertFalse("Expected no error, but an error was returned", resultOrErr.hasGerr());
-        Assert.assertEquals(expectedResult, resultOrErr.getVal().longValue());
+        Assert.assertEquals(expectedResult, resultOrErr.getVal());
     }
 
     @Test
     public void testSubtract() {
-        var oldValue = 10L;
-        var toSubtract = 5L;
-        var expectedResult = oldValue - toSubtract;
+        var oldValue = new BigDecimal(10);
+        var toSubtract = new BigDecimal(5);
+        var expectedResult = oldValue.subtract(toSubtract);
         var resultOrErr = AssetParameterChangeValueOperation.SUBTRACT.apply(oldValue, toSubtract);
         Assert.assertFalse("Expected no error, but an error was returned", resultOrErr.hasGerr());
-        Assert.assertEquals(expectedResult, resultOrErr.getVal().longValue());
+        Assert.assertEquals(expectedResult, resultOrErr.getVal());
     }
 
     @Test
     public void testOversubtractionError() {
-        var oldValue = 5L;
-        var toSubtract = 10L;
-        var expectedResult = oldValue - toSubtract;
+        var oldValue = new BigDecimal(5);
+        var toSubtract = new BigDecimal(10);
         var resultOrErr = AssetParameterChangeValueOperation.SUBTRACT.apply(oldValue, toSubtract);
         Assert.assertTrue("Expected an error on oversubtraction", resultOrErr.hasGerr());
     }
 
     @Test
     public void testSet() {
-        var oldValue = 5L;
-        var toSet = 100L;
-        var resultOrErr = AssetParameterChangeValueOperation.SUBTRACT.apply(oldValue, toSet);
+        var oldValue = new BigDecimal(5);
+        var toSet = new BigDecimal(100);
+        var resultOrErr = AssetParameterChangeValueOperation.SET.apply(oldValue, toSet);
         Assert.assertFalse("Expected no error, but an error was returned", resultOrErr.hasGerr());
-        Assert.assertEquals(toSet, resultOrErr.getVal().longValue());
+        Assert.assertEquals(toSet, resultOrErr.getVal());
     }
 }