+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Split net worth rendering into its own class to control complexity
 * Refactor the projections calculator to use the new asset datamodel
 * Update to Java 11 to get the lovely `var` keyword
+* Switch to using BigDecimal for all networth calculations
 
 ## 0.5.0
 * Allow users to base scenarios on other scenarios